Simulated Dataset with 1D Geography

A simualted dataset to show the utility of this package

admixed

Format

A list with the following elements

x

simulated design matrix, 500 x 200 matrix

causal

character vector of the names of the causal SNPs

beta

the vector of true regression coefficients. 5 are causal

kin

2 times the estimated kinship

Xkinship

the matrix of SNPs used to estimate the kinship matrix, 500 x 10,000 matirx

not_causal

character vector of the non-causal SNPs

causal_positive

character vector of the causal SNPs with positive regression coefficient

causal_negative

character vector of the causal SNPs with negative regression coefficient

x_lasso

the design matrix which also includes nPC principal components

Details

The code used to simulate the data is available at https://github.com/sahirbhatnagar/ggmix/blob/master/data-raw/bnpsd-data.R. See gen_structured_model for more details on the output and how the function used to simulate the data.

References

Ochoa, Alejandro, and John D. Storey. 2016a. "FST And Kinship for Arbitrary Population Structures I: Generalized Definitions." bioRxiv doi:10.1101/083915.

Ochoa, Alejandro, and John D. Storey. 2016b. "FST And Kinship for Arbitrary Population Structures II: Method of Moments Estimators." bioRxiv doi:10.1101/083923.

Examples

data(admixed)
str(admixed)

#> List of 10
#>  $ y              : num [1:500] -1.382 -0.203 -1.806 -1.858 -0.522 ...
#>  $ x              : int [1:500, 1:200] 2 2 1 0 0 0 1 1 0 1 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:500] "id1" "id2" "id3" "id4" ...
#>   .. ..$ : chr [1:200] "X62" "X185" "X371" "X420" ...
#>  $ causal         : chr [1:10] "X336" "X7638" "X1199" "X1536" ...
#>  $ beta           : num [1:200] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ kin            : num [1:500, 1:500] 1.071 0.185 0.173 0.174 0.179 ...
#>  $ Xkinship       : int [1:500, 1:10000] 0 0 0 0 0 0 0 0 1 0 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:500] "id1" "id2" "id3" "id4" ...
#>   .. ..$ : chr [1:10000] "X8423" "X10014" "X3265" "X8768" ...
#>  $ not_causal     : chr [1:190] "X62" "X185" "X371" "X420" ...
#>  $ causal_positive: chr [1:4] "X7638" "X1199" "X1536" "X1943"
#>  $ causal_negative: chr [1:6] "X336" "X1968" "X2849" "X8087" ...
#>  $ x_lasso        : num [1:500, 1:210] 2 2 1 0 0 0 1 1 0 1 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:500] "id1" "id2" "id3" "id4" ...
#>   .. ..$ : chr [1:210] "X62" "X185" "X371" "X420" ...