Simulation Scenarion from Bhatnagar et al. (2018+) ggmix paper

Function that generates data of the different simulation studies presented in the accompanying paper. This function requires the popkin and bnpsd package to be installed.

gen_structured_model(n, p_test, p_kinship, k, s, Fst, b0, nPC = 10, eta,
  sigma2, geography = c("ind", "1d", "circ"), percent_causal, percent_overlap)

Arguments

n
number of observations to simulate

p_test
number of variables in X_test, i.e., the design matrix

p_kinship
number of variable in X_kinship, i.e., matrix used to calculate kinship

k
number of intermediate subpopulations.

s
the desired bias coefficient, which specifies sigma indirectly. Required if sigma is missing

Fst
The desired final FST of the admixed individuals. Required if sigma is missing

b0
the true intercept parameter

nPC
number of principal components to include in the design matrix used for regression adjustment for population structure via principal components. This matrix is used as the input in a standard lasso regression routine, where there are no random effects.

eta
the true eta parameter, which has to be 0 < eta < 1

sigma2
the true sigma2 parameter

geography
the type of geography for simulation the kinship matrix. "ind" is independent populations where every individuals is actually unadmixed, "1d" is a 1D geography and "circ" is circular geography. Default: "ind". See the functions in the bnpsd for details on how this data is actually generated.

percent_causal
percentage of p_test that is causal. must be \(0 \leq percent_causal \leq 1\). The true regression coefficients are generated from a standard normal distribution.

percent_overlap
this represents the percentage of causal SNPs that will also be included in the calculation of the kinship matrix

Value

A list with the following elements

x

simulated design matrix

causal

character vector of the names of the causal SNPs

beta

the vector of true regression coefficients

kin

2 times the estimated kinship

Xkinship

the matrix of SNPs used to estimate the kinship matrix

not_causal

character vector of the non-causal SNPs

causal_positive

character vector of the causal SNPs with positive regression coefficient

causal_negative

character vector of the causal SNPs with negative regression coefficient

x_lasso

the design matrix which also includes nPC principal components

Details

The kinship is estimated using the popkin function from the popkin package. This function will multiple that kinship matrix by 2 to give the expected covariance matrix which is subsequently used in the linear mixed models

See also