Generalised Information Criterion

Calculates the generalised information criterion for each value of the tuning parameter lambda

gic(ggmix_fit, ...)

# S3 method for default
gic(ggmix_fit, ...)

# S3 method for ggmix_fit
gic(ggmix_fit, ..., an = log(log(n)) * log(p))

Arguments

ggmix_fit
An object of class ggmix_fit which is outputted by the ggmix function

...
other parameters. currently ignored.

an
numeric, the penalty per parameter to be used; the default is an = log(log(n))*log(p) where n is the number of subjects and p is the number of parameters

Value

an object with S3 class "ggmix_gic", "ggmix_fit", "*" and "**" where "*" is "lasso" or "gglasso" and "**" is fullrank or lowrank. Results are provided for converged values of lambda only.

lambda

the sequence of converged tuning parameters

nzero

the number of non-zero estimated coefficients including the 2 variance parameters which are not penalized and therefore always included

gic

gic value. a numeric vector with length equal to length(lambda)

lambda.min.name

a character corresponding to the name of the tuning parameter lambda which minimizes the gic

lambda.min

the value of lambda which minimizes the gic

Details

the generalised information criterion used for gaussian response is given by $$-2 * loglikelihood(\hat{\Theta}) + an * df$$ where df is the number of non-zero estimated parameters, including variance components

References

Fan Y, Tang CY. Tuning parameter selection in high dimensional penalized likelihood. Journal of the Royal Statistical Society: Series B (Statistical Methodology). 2013 Jun 1;75(3):531-52.

Nishii R. Asymptotic properties of criteria for selection of variables in multiple regression. The Annals of Statistics. 1984;12(2):758-65.

See also