Estimation of Log-likelihood for Linear Mixed Model with Lasso Penalty

sigma2lasso estimates the value of the sigma2 for the linear mixed model with lasso penalty

logliklasso(ggmix_object, ...)

# S3 method for default
logliklasso(ggmix_object, ...)

# S3 method for fullrank
logliklasso(ggmix_object, ..., eta, sigma2, beta, nt, x, y)

Arguments

ggmix_object
A ggmix_object object of class lowrank or fullrank

...
Extra parameters. Currently ignored.

eta
current estimate of the eta parameter

sigma2
current estimate of the sigma2 parameter

beta
current estimate of the beta parameter including the intercept. this should be of length p+1, where p is the number of variables.

nt
total number of observations

x
input matrix, of dimension n x p; where n is the number of observations and p are the number of predictors.

y
response variable. must be a quantitative variable

Value

A decreasing sequence of tuning parameters

Note

This function isn't meant to be called directly by the user.

See also