Extract Random Effects

Generic function for extracting the random effects. This is the same generic (and same name) defined in the nlme and lme4 package.

ranef(object, ...)

random.effects(object, ...)

# S3 method for default
random.effects(object, ...)

# S3 method for default
ranef(object, ...)

# S3 method for ggmix_gic
ranef(object, s = "lambda.min", ...)

Arguments

object
any fitted model object from which random effects estimates can be extracted. Currently supports "ggmix_gic" objects outputted by the gic function

...
other parameters. currently ignored

s
Value(s) of the penalty parameter lambda at which predictions are required. Default is the value s="lambda.min" can be used. If s is numeric, it is taken as the value(s) of lambda to be used.

Value

a numeric vector of length equal to the number of observations of subject-specific random effects

Details

For objects of class "ggmix_gic", this function returns the subject-specific random effect value for the model which minimizes the GIC using the maximum a posteriori principle

See also