Simulated Dataset with 1D Geography

A simulated dataset to show the utility of this package

admixed

Format

An object of class list of length 21.

Value

A list with the following elements

ytune

simulated response vector for tuning parameter selection set

ytest

simulated response vector for test set

xtrain

simulated design matrix for training set

xtune

simulated design matrix for tuning parameter selection set

xtest

simulated design matrix for testing set

xtrain_lasso

simulated design matrix for training set for lasso model. This is the same as xtrain, but also includes the nPC principal components

xtune_lasso

simulated design matrix for tuning parameter selection set for lasso model. This is the same as xtune, but also includes the nPC principal components

xtest

simulated design matrix for testing set for lasso model. This is the same as xtest, but also includes the nPC principal components

causal

character vector of the names of the causal SNPs

beta

the vector of true regression coefficients

kin_train

2 times the estimated kinship for the training set individuals

kin_tune_train

The covariance matrix between the tuning set and the training set individuals

kin_test_train

The covariance matrix between the test set and training set individuals

Xkinship

the matrix of SNPs used to estimate the kinship matrix

not_causal

character vector of the non-causal SNPs

PC

the principal components for population structure adjustment

Details

The code used to simulate the data is available at https://github.com/sahirbhatnagar/ggmix/blob/master/data-raw/bnpsd-data.R. See gen_structured_model for more details on the output and how the function used to simulate the data.

References

Ochoa, Alejandro, and John D. Storey. 2016a. "FST And Kinship for Arbitrary Population Structures I: Generalized Definitions." bioRxiv doi:10.1101/083915.

Ochoa, Alejandro, and John D. Storey. 2016b. "FST And Kinship for Arbitrary Population Structures II: Method of Moments Estimators." bioRxiv doi:10.1101/083923.

Examples

data(admixed)
str(admixed)

#> List of 21
#>  $ ytrain        : Named num [1:80] 1.78783 -0.00688 -0.66998 -1.6918 -0.22518 ...
#>   ..- attr(*, "names")= chr [1:80] "id1" "id2" "id3" "id4" ...
#>  $ ytune         : Named num [1:10] -3.337 -1.642 -0.494 2.487 2.57 ...
#>   ..- attr(*, "names")= chr [1:10] "id21" "id25" "id28" "id49" ...
#>  $ ytest         : Named num [1:10] 3.123 -0.244 -1.608 -1.396 1.077 ...
#>   ..- attr(*, "names")= chr [1:10] "id26" "id39" "id45" "id52" ...
#>  $ xtrain        : int [1:80, 1:50] 0 0 1 0 2 2 0 0 0 0 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:80] "id1" "id2" "id3" "id4" ...
#>   .. ..$ : chr [1:50] "X23" "X36" "X38" "X40" ...
#>  $ xtune         : int [1:10, 1:50] 0 0 1 1 1 1 1 1 0 1 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:10] "id21" "id25" "id28" "id49" ...
#>   .. ..$ : chr [1:50] "X23" "X36" "X38" "X40" ...
#>  $ xtest         : int [1:10, 1:50] 0 1 1 1 1 1 2 0 2 2 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:10] "id26" "id39" "id45" "id52" ...
#>   .. ..$ : chr [1:50] "X23" "X36" "X38" "X40" ...
#>  $ xtrain_lasso  : num [1:80, 1:60] 0 0 1 0 2 2 0 0 0 0 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:80] "id1" "id2" "id3" "id4" ...
#>   .. ..$ : chr [1:60] "X23" "X36" "X38" "X40" ...
#>  $ xtune_lasso   : num [1:10, 1:60] 0 0 1 1 1 1 1 1 0 1 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:10] "id21" "id25" "id28" "id49" ...
#>   .. ..$ : chr [1:60] "X23" "X36" "X38" "X40" ...
#>  $ xtest_lasso   : num [1:10, 1:60] 0 1 1 1 1 1 2 0 2 2 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:10] "id26" "id39" "id45" "id52" ...
#>   .. ..$ : chr [1:60] "X23" "X36" "X38" "X40" ...
#>  $ Xkinship      : int [1:80, 1:500] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:80] "id1" "id2" "id3" "id4" ...
#>   .. ..$ : chr [1:500] "X279" "X295" "X346" "X304" ...
#>  $ kin_train     : num [1:80, 1:80] 1.0303 0.0722 0.0664 0.0315 0.1244 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:80] "id1" "id2" "id3" "id4" ...
#>   .. ..$ : chr [1:80] "id1" "id2" "id3" "id4" ...
#>  $ kin_tune_train: num [1:10, 1:80] 0.08378 0.03152 0.00248 -0.02074 0.00829 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:10] "id21" "id25" "id28" "id49" ...
#>   .. ..$ : chr [1:80] "id1" "id2" "id3" "id4" ...
#>  $ kin_test_train: num [1:10, 1:80] 0.078 0.0257 0.1244 0.0664 -0.0324 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:10] "id26" "id39" "id45" "id52" ...
#>   .. ..$ : chr [1:80] "id1" "id2" "id3" "id4" ...
#>  $ mu_train      : num [1:80] 1.81 1.84 -1.45 -1.24 0.86 ...
#>  $ causal        : chr [1:5] "X407" "X507" "X524" "X538" ...
#>  $ beta          : num [1:50] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ not_causal    : chr [1:45] "X23" "X36" "X38" "X40" ...
#>  $ kinship       : num [1:100, 1:100] 0.5281 0.0557 0.0553 0.0549 0.0544 ...
#>  $ coancestry    : num [1:100, 1:100] 0.0561 0.0557 0.0553 0.0549 0.0544 ...
#>  $ PC            : num [1:100, 1:10] 0.855 -1.15 -0.514 -0.443 -1.219 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:100] "id1" "id2" "id3" "id4" ...
#>   .. ..$ : chr [1:10] "PC1" "PC2" "PC3" "PC4" ...
#>  $ subpops       : num [1:100] 1 1 1 1 1 1 1 1 1 1 ...